refactor whole back-end
